requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
PyJWT==2.9.0
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
